[ML-82] Refactor tests
